use font-based units for column widths
This way all layout sizing is derived from the space inner-most
elements (table cells in our case) require to display their data,
outwards through all containers.

The idea is to minimize reliance on units that are not portable across
displays (most commonly px).

For this to work it is important to place reliable restrictions on
those inner-most elements that define all sizing.