fix unhandled errors when auto-updating markets
Silencing network errors in markets auto-update loop is fair because
next update must be scheduled regardless of the success of the current
update.

This change fixes printing of two unhandled promise rejections in
fetchMarketsLoop. One is when toggleMarketsUpdating is aborting a
running fetch. Another is when both a user action and update loop timer
request two concurrent fetches and one of them gets rejected.