explicitly store and validate selected market id
Change asyncUpdateSelectedBooks function from getting data from
marketsTable to using an explicitly stored market id value. This way it
doesn't need to know about UI structure (and UI at all) to start
fetching of the books.

This is a step to maintaining business data regardless of the UI.

Two other changes while here:

- extract functions to test and convert to a valid market id
- update market selection UI state only if the new id is valid

The latter may help to spot a bug if market selection does not change
as expected.