move WS markets consistency code to model section
After UI knowledge has been removed from `connect()`, what remains in it
is the consistency code ensuring the markets model is populated before
subscribing to WebSocket API updates.

For UI code it is an implementation detail of the model that UI must not
be concerned about. Also, UI code must not use Poloniex API directly.