allow to unsubscribe from ticker updates
A workaround was required. Poloniex WS API returns channel id as integer
to confirm subscription: `[1002, 1]`. But when confirming
un-subscription it returns a string: `["1002", 0]`.

While here, correct comment in the `default` case. It was not "data we
didn't subscribe for" (there is no stateful subscription management at
the moment) but rather "data of unknown type".